Switch to harmony2.0
Switch to harmony2.0
