+++ CustomAI/RealCityResidentAI.cs
@@ -1,14 +1,10 @@
 ﻿using System;
 using ColossalFramework;
-using UnityEngine;
 using RealCity.Util;
 using RealCity.UI;
-using Harmony;
-using System.Reflection;
 
 namespace RealCity.CustomAI
 {
-    [HarmonyPatch]
     public static class RealCityResidentAI
     {
         public static uint preCitizenId = 0;
@@ -52,11 +48,6 @@ namespace RealCity.CustomAI
 
         public static byte[] saveData = new byte[144];
         //public static byte[] saveData = new byte[140];
-
-        public static MethodBase TargetMethod()
-        {
-            return typeof(ResidentAI).GetMethod("SimulationStep", BindingFlags.Public | BindingFlags.Instance, null, new Type[] { typeof(uint), typeof(CitizenUnit).MakeByRefType() }, null);
-        }
 
         public static void Load()
         {
@@ -499,766 +490,6 @@ namespace RealCity.CustomAI
                 MainDataStore.building_buffer1[buildingID] = totalWorkCount;
             }
             return totalWorkCount;
-        }
-
-        public static void ProcessCitizen(uint homeID, ref CitizenUnit data, bool isPre)
-        {
-            if (isPre)
-            {
-                MainDataStore.family_money[homeID] = 0;
-                if (data.m_citizen0 != 0)
-                {
-                    Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen0];
-                    if ((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None)
-                    {
-                        if (citizenData.Dead == false)
-                        {
-                            citizenCount++;
-                            MainDataStore.family_money[homeID] += MainDataStore.citizenMoney[data.m_citizen0];
-                        }
-                    }
-                }
-                if (data.m_citizen1 != 0)
-                {
-                    Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen1];
-                    if ((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None)
-                    {
-                        if (citizenData.Dead == false)
-                        {
-                            citizenCount++;
-                            MainDataStore.family_money[homeID] += MainDataStore.citizenMoney[data.m_citizen1];
-                        }
-                    }
-                }
-                if (data.m_citizen2 != 0)
-                {
-                    Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen2];
-                    if ((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None)
-                    {
-                        if (citizenData.Dead == false)
-                        {
-                            citizenCount++;
-                            MainDataStore.family_money[homeID] += MainDataStore.citizenMoney[data.m_citizen2];
-                        }
-                    }
-                }
-                if (data.m_citizen3 != 0)
-                {
-                    Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen3];
-                    if ((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None)
-                    {
-                        if (citizenData.Dead == false)
-                        {
-                            citizenCount++;
-                            MainDataStore.family_money[homeID] += MainDataStore.citizenMoney[data.m_citizen3];
-                        }
-                    }
-                }
-                if (data.m_citizen4 != 0)
-                {
-                    Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen4];
-                    if ((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None)
-                    {
-                        if (citizenData.Dead == false)
-                        {
-                            citizenCount++;
-                            MainDataStore.family_money[homeID] += MainDataStore.citizenMoney[data.m_citizen4];
-                        }
-                    }
-                }
-            }
-            else
-            {
-                if (MainDataStore.family_money[homeID] < 5000)
-                {
-                    familyWeightStableLow = (ushort)(familyWeightStableLow + 1);
-                }
-                else if (MainDataStore.family_money[homeID] >= 20000)
-                {
-                    familyWeightStableHigh = (ushort)(familyWeightStableHigh + 1);
-                }
-
-                int temp = 0;
-                if (data.m_citizen0 != 0)
-                {
-                    Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen0];
-                    if (((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None) && (citizenData.Dead == false))
-                    {
-                        temp++;
-#if FASTRUN
-#else
-                        GetVoteChance(data.m_citizen0, citizenData, homeID);
-#endif
-                    }
-                }
-                if (data.m_citizen1 != 0)
-                {
-                    Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen1];
-                    if (((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None) && (citizenData.Dead == false))
-                    {
-#if FASTRUN
-#else
-                        GetVoteChance(data.m_citizen1, citizenData, homeID);
-#endif
-                        temp++;
-                    }
-                }
-                if (data.m_citizen2 != 0)
-                {
-                    Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen2];
-                    if (((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None) && (citizenData.Dead == false))
-                    {
-#if FASTRUN
-#else
-                        GetVoteChance(data.m_citizen2, citizenData, homeID);
-#endif
-                        temp++;
-                    }
-                }
-                if (data.m_citizen3 != 0)
-                {
-                    Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen3];
-                    if (((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None) && (citizenData.Dead == false))
-                    {
-#if FASTRUN
-#else
-                        GetVoteChance(data.m_citizen3, citizenData, homeID);
-#endif
-                        temp++;
-                    }
-                }
-                if (data.m_citizen4 != 0)
-                {
-                    Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen4];
-                    if (((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None) && (citizenData.Dead == false))
-                    {
-#if FASTRUN
-#else
-                        GetVoteChance(data.m_citizen4, citizenData, homeID);
-#endif
-                        temp++;
-                    }
-                }
-
-                if (temp != 0)
-                {
-                    if (data.m_citizen0 != 0)
-                    {
-                        Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen0];
-                        if (((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None) && (citizenData.Dead == false))
-                        {
-                            MainDataStore.citizenMoney[data.m_citizen0] = MainDataStore.family_money[homeID] / temp;
-                        }
-                    }
-                    if (data.m_citizen1 != 0)
-                    {
-                        Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen1];
-                        if (((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None) && (citizenData.Dead == false))
-                        {
-                            MainDataStore.citizenMoney[data.m_citizen1] = MainDataStore.family_money[homeID] / temp;
-                        }
-                    }
-                    if (data.m_citizen2 != 0)
-                    {
-                        Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen2];
-                        if (((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None) && (citizenData.Dead == false))
-                        {
-                            MainDataStore.citizenMoney[data.m_citizen2] = MainDataStore.family_money[homeID] / temp;
-                        }
-                    }
-                    if (data.m_citizen3 != 0)
-                    {
-                        Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen3];
-                        if (((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None) && (citizenData.Dead == false))
-                        {
-                            MainDataStore.citizenMoney[data.m_citizen3] = MainDataStore.family_money[homeID] / temp;
-                        }
-                    }
-                    if (data.m_citizen4 != 0)
-                    {
-                        Citizen citizenData = Singleton<CitizenManager>.instance.m_citizens.m_buffer[data.m_citizen4];
-                        if (((citizenData.m_flags & Citizen.Flags.MovingIn) == Citizen.Flags.None) && (citizenData.Dead == false))
-                        {
-                            MainDataStore.citizenMoney[data.m_citizen4] = MainDataStore.family_money[homeID] / temp;
-                        }
-                    }
-                }
-            }
-
-        }
-
-        public static byte ProcessFamily(uint homeID, ref CitizenUnit data)
-        {
-            if (preCitizenId > homeID)
-            {
-                MainDataStore.familyCount = familyCount;
-                MainDataStore.citizenCount = citizenCount;
-                MainDataStore.family_profit_money_num = familyProfitMoneyCount;
-                MainDataStore.family_very_profit_money_num = familyVeryProfitMoneyCount;
-                MainDataStore.family_loss_money_num = familyLossMoneyCount;
-                if (familyCount != 0)
-                {
-                    MainDataStore.citizenSalaryPerFamily = ((citizenSalaryCount / familyCount));
-                    MainDataStore.citizenExpensePerFamily = ((citizenExpenseCount / familyCount));
-                }
-                MainDataStore.citizenExpense = citizenExpenseCount;
-                MainDataStore.citizenSalaryTaxTotal = citizenSalaryTaxTotal;
-                MainDataStore.citizenSalaryTotal = citizenSalaryCount;
-                if (MainDataStore.familyCount < MainDataStore.family_weight_stable_high)
-                {
-                    MainDataStore.family_weight_stable_high = (uint)MainDataStore.familyCount;
-                }
-                else
-                {
-                    MainDataStore.family_weight_stable_high = familyWeightStableHigh;
-                }
-                if (MainDataStore.familyCount < MainDataStore.family_weight_stable_low)
-                {
-                    MainDataStore.family_weight_stable_low = (uint)MainDataStore.familyCount;
-                }
-                else
-                {
-                    MainDataStore.family_weight_stable_low = familyWeightStableLow;
-                }
-                citizenGoods = citizenGoodsTemp;
-                familyVeryProfitMoneyCount = 0;
-                familyProfitMoneyCount = 0;
-                familyLossMoneyCount = 0;
-                familyCount = 0;
-                citizenCount = 0;
-                citizenSalaryCount = 0;
-                citizenExpenseCount = 0;
-                citizenSalaryTaxTotal = 0;
-                tempCitizenSalaryTaxTotal = 0f;
-                familyWeightStableHigh = 0;
-                familyWeightStableLow = 0;
-                citizenGoodsTemp = 0;
-            }
-            preCitizenId = homeID;
-
-            familyCount++;
-            citizenGoodsTemp += data.m_goods;
-
-            if (homeID > 524288)
-            {
-                DebugLog.LogToFileOnly("Error: citizen ID greater than 524288");
-            }
-            //ProcessCitizen pre, gather all citizenMoney to family
-            ProcessCitizen(homeID, ref data, true);
-            //1.We caculate citizen income
-            int familySalaryCurrent = 0;
-            familySalaryCurrent += ProcessCitizenSalary(data.m_citizen0, false);
-            familySalaryCurrent += ProcessCitizenSalary(data.m_citizen1, false);
-            familySalaryCurrent += ProcessCitizenSalary(data.m_citizen2, false);
-            familySalaryCurrent += ProcessCitizenSalary(data.m_citizen3, false);
-            familySalaryCurrent += ProcessCitizenSalary(data.m_citizen4, false);
-            citizenSalaryCount = citizenSalaryCount + familySalaryCurrent;
-            if (familySalaryCurrent < 0)
-            {
-                DebugLog.LogToFileOnly("familySalaryCurrent< 0 in ResidentAI");
-                familySalaryCurrent = 0;
-            }
-
-            //2.We caculate salary tax
-            float tax = (float)Politics.residentTax * familySalaryCurrent / 100f;
-            tempCitizenSalaryTaxTotal = tempCitizenSalaryTaxTotal + (int)tax;
-            citizenSalaryTaxTotal = (int)tempCitizenSalaryTaxTotal;
-            ProcessCitizenIncomeTax(homeID, tax);
-
-            //3. We caculate expense
-            int educationFee = 0;
-            int expenseRate = 0;
-            CitizenManager instance = Singleton<CitizenManager>.instance;
-            if (data.m_citizen4 != 0u && !instance.m_citizens.m_buffer[(int)((UIntPtr)data.m_citizen4)].Dead)
-            {
-                expenseRate = 0;
-                educationFee += GetExpenseRate(homeID, data.m_citizen4, out expenseRate);
-            }
-            if (data.m_citizen3 != 0u && !instance.m_citizens.m_buffer[(int)((UIntPtr)data.m_citizen3)].Dead)
-            {
-                expenseRate = 0;
-                educationFee += GetExpenseRate(homeID, data.m_citizen3, out expenseRate);
-            }
-            if (data.m_citizen2 != 0u && !instance.m_citizens.m_buffer[(int)((UIntPtr)data.m_citizen2)].Dead)
-            {
-                expenseRate = 0;
-                educationFee += GetExpenseRate(homeID, data.m_citizen2, out expenseRate);
-            }
-            if (data.m_citizen1 != 0u && !instance.m_citizens.m_buffer[(int)((UIntPtr)data.m_citizen1)].Dead)
-            {
-                expenseRate = 0;
-                educationFee += GetExpenseRate(homeID, data.m_citizen1, out expenseRate);
-            }
-            if (data.m_citizen0 != 0u && !instance.m_citizens.m_buffer[(int)((UIntPtr)data.m_citizen0)].Dead)
-            {
-                expenseRate = 0;
-                educationFee += GetExpenseRate(homeID, data.m_citizen0, out expenseRate);
-            }
-            ProcessCitizenHouseRent(homeID, expenseRate);
-            //campus DLC added.
-            expenseRate = UniqueFacultyAI.IncreaseByBonus(UniqueFacultyAI.FacultyBonus.Economics, expenseRate);
-            citizenExpenseCount += (educationFee + expenseRate);
-
-            //4. income - expense
-            float incomeMinusExpense = familySalaryCurrent - tax - educationFee - expenseRate;
-            MainDataStore.family_money[homeID] += incomeMinusExpense;
-
-            //5. Process shopping
-            if (MainDataStore.familyGoods[homeID] == 0)
-            {
-                //first time
-            }
-            else if (MainDataStore.familyGoods[homeID] < data.m_goods)
-            {
-                MainDataStore.family_money[homeID] -= RealCityIndustryBuildingAI.GetResourcePrice(TransferManager.TransferReason.Shopping) * (data.m_goods - MainDataStore.familyGoods[homeID]);
-            }
-
-            //6. Process citizen status
-            if (incomeMinusExpense <= 0)
-            {
-                familyLossMoneyCount = (familyLossMoneyCount + 1);
-            }
-            else if (incomeMinusExpense > 100)
-            {
-                familyVeryProfitMoneyCount = (familyVeryProfitMoneyCount + 1);
-            }
-            else
-            {
-                familyProfitMoneyCount = (familyProfitMoneyCount + 1);
-            }
-
-            //7. Caculate minimumLivingAllowance and benefitOffset
-            if (MainDataStore.family_money[homeID] < (-Politics.benefitOffset))
-            {
-                int num = (int)(-(MainDataStore.family_money[homeID]) + 0.5f + Politics.benefitOffset);
-                MainDataStore.family_money[homeID] = 0;
-                MainDataStore.minimumLivingAllowance += num;
-                Singleton<EconomyManager>.instance.FetchResource((EconomyManager.Resource)17, num, ItemClass.Service.Residential, ItemClass.SubService.None, ItemClass.Level.Level1);
-            }
-            else
-            {
-                if (Politics.benefitOffset > 0)
-                {
-                    MainDataStore.family_money[homeID] += Politics.benefitOffset;
-                    MainDataStore.minimumLivingAllowance += Politics.benefitOffset;
-                    Singleton<EconomyManager>.instance.FetchResource((EconomyManager.Resource)17, Politics.benefitOffset, ItemClass.Service.Residential, ItemClass.SubService.None, ItemClass.Level.Level1);
-                }
-            }
-
-            //8. Limit familyMoney
-            if (MainDataStore.family_money[homeID] > 32000000f)
-            {
-                MainDataStore.family_money[homeID] = 32000000f;
-            }
-
-            if (MainDataStore.family_money[homeID] < -32000000f)
-            {
-                MainDataStore.family_money[homeID] = -32000000f;
-            }
-
-            //ProcessCitizen post, split all familyMoney to citizen
-            ProcessCitizen(homeID, ref data, false);
-            //9. Fixed m_goods consuption
-            //9.1 based on incomeMinusExpense
-            float fixedGoodsConsumption = incomeMinusExpense / 10f;
-            //9.2 based on familyMoney
-            if (MainDataStore.family_money[homeID] > 0)
-            {
-                fixedGoodsConsumption += (int)(MainDataStore.family_money[homeID] / 5000);
-            }
-            if (fixedGoodsConsumption < 1)
-            {
-                fixedGoodsConsumption = 1;
-            }
-            else if (fixedGoodsConsumption > 20)
-            {
-                fixedGoodsConsumption = 20;
-            }
-            return (byte)fixedGoodsConsumption;
-        }
-
-
-        public static void ProcessCitizenIncomeTax(uint homeID, float tax)
-        {
-            CitizenManager instance = Singleton<CitizenManager>.instance;
-            ushort building = instance.m_units.m_buffer[(int)((UIntPtr)homeID)].m_building;
-            Building buildingdata = Singleton<BuildingManager>.instance.m_buildings.m_buffer[building];
-            Singleton<EconomyManager>.instance.AddPrivateIncome((int)(tax), buildingdata.Info.m_class.m_service, buildingdata.Info.m_class.m_subService, buildingdata.Info.m_class.m_level, 112333);
-        }
-
-        public static void ProcessCitizenHouseRent(uint homeID, int expenserate)
-        {
-            CitizenManager instance = Singleton<CitizenManager>.instance;
-            ushort building = instance.m_units.m_buffer[(int)((UIntPtr)homeID)].m_building;
-            Building buildingdata = Singleton<BuildingManager>.instance.m_buildings.m_buffer[building];
-            DistrictManager instance2 = Singleton<DistrictManager>.instance;
-            byte district = instance2.GetDistrict(buildingdata.m_position);
-            Singleton<EconomyManager>.instance.AddPrivateIncome(expenserate * 100, buildingdata.Info.m_class.m_service, buildingdata.Info.m_class.m_subService, buildingdata.Info.m_class.m_level, 100);
-        }
-
-        // ResidentAI
-        public static void Prefix(uint homeID, ref CitizenUnit data)
-        {
-            // NON-STOCK CODE START
-            // -20 because later in game code, m_goods will minus 20
-            int fixedGoodsConsumption = 0;
-            if ((Singleton<BuildingManager>.instance.m_buildings.m_buffer[data.m_building].m_flags & (Building.Flags.Completed | Building.Flags.Upgrading)) != Building.Flags.None)
-            {
-                fixedGoodsConsumption = ProcessFamily(homeID, ref data) - 20;
-            }
-            data.m_goods = (ushort)Mathf.Max(21, (data.m_goods - fixedGoodsConsumption)); //here we can adjust demand
-            MainDataStore.familyGoods[homeID] = (ushort)Mathf.Max(0, data.m_goods - 20);
-            // NON-STOCK CODE END
-        }
-
-        public static int GetExpenseRate(uint homeid, uint citizenID, out int incomeAccumulation)
-        {
-            BuildingManager instance1 = Singleton<BuildingManager>.instance;
-            CitizenManager instance2 = Singleton<CitizenManager>.instance;
-            ItemClass @class = instance1.m_buildings.m_buffer[instance2.m_citizens.m_buffer[citizenID].m_homeBuilding].Info.m_class;
-            incomeAccumulation = 0;
-            DistrictManager instance = Singleton<DistrictManager>.instance;
-            if (instance2.m_citizens.m_buffer[citizenID].m_homeBuilding != 0)
-            {
-                byte district = instance.GetDistrict(instance1.m_buildings.m_buffer[instance2.m_citizens.m_buffer[citizenID].m_homeBuilding].m_position);
-                DistrictPolicies.Taxation taxationPolicies = instance.m_districts.m_buffer[district].m_taxationPolicies;
-                if (@class.m_subService == ItemClass.SubService.ResidentialLow)
-                {
-                    switch (@class.m_level)
-                    {
-                        case ItemClass.Level.Level1:
-                            incomeAccumulation = MainDataStore.residentLowLevel1Rent;
-                            break;
-                        case ItemClass.Level.Level2:
-                            incomeAccumulation = MainDataStore.residentLowLevel2Rent;
-                            break;
-                        case ItemClass.Level.Level3:
-                            incomeAccumulation = MainDataStore.residentLowLevel3Rent;
-                            break;
-                        case ItemClass.Level.Level4:
-                            incomeAccumulation = MainDataStore.residentLowLevel4Rent;
-                            break;
-                        case ItemClass.Level.Level5:
-                            incomeAccumulation = MainDataStore.residentLowLevel5Rent;
-                            break;
-                    }
-                }
-                else if (@class.m_subService == ItemClass.SubService.ResidentialLowEco)
-                {
-                    switch (@class.m_level)
-                    {
-                        case ItemClass.Level.Level1:
-                            incomeAccumulation = MainDataStore.residentLowLevel1Rent << 1;
-                            break;
-                        case ItemClass.Level.Level2:
-                            incomeAccumulation = MainDataStore.residentLowLevel2Rent << 1;
-                            break;
-                        case ItemClass.Level.Level3:
-                            incomeAccumulation = MainDataStore.residentLowLevel3Rent << 1;
-                            break;
-                        case ItemClass.Level.Level4:
-                            incomeAccumulation = MainDataStore.residentLowLevel4Rent << 1;
-                            break;
-                        case ItemClass.Level.Level5:
-                            incomeAccumulation = MainDataStore.residentLowLevel5Rent << 1;
-                            break;
-                    }
-                }
-                else if (@class.m_subService == ItemClass.SubService.ResidentialHigh)
-                {
-                    switch (@class.m_level)
-                    {
-                        case ItemClass.Level.Level1:
-                            incomeAccumulation = MainDataStore.residentHighLevel1Rent;
-                            break;
-                        case ItemClass.Level.Level2:
-                            incomeAccumulation = MainDataStore.residentHighLevel2Rent;
-                            break;
-                        case ItemClass.Level.Level3:
-                            incomeAccumulation = MainDataStore.residentHighLevel3Rent;
-                            break;
-                        case ItemClass.Level.Level4:
-                            incomeAccumulation = MainDataStore.residentHighLevel4Rent;
-                            break;
-                        case ItemClass.Level.Level5:
-                            incomeAccumulation = MainDataStore.residentHighLevel5Rent;
-                            break;
-                    }
-                }
-                else
-                {
-                    switch (@class.m_level)
-                    {
-                        case ItemClass.Level.Level1:
-                            incomeAccumulation = MainDataStore.residentHighLevel1Rent << 1;
-                            break;
-                        case ItemClass.Level.Level2:
-                            incomeAccumulation = MainDataStore.residentHighLevel2Rent << 1;
-                            break;
-                        case ItemClass.Level.Level3:
-                            incomeAccumulation = MainDataStore.residentHighLevel3Rent << 1;
-                            break;
-                        case ItemClass.Level.Level4:
-                            incomeAccumulation = MainDataStore.residentHighLevel4Rent << 1;
-                            break;
-                        case ItemClass.Level.Level5:
-                            incomeAccumulation = MainDataStore.residentHighLevel5Rent << 1;
-                            break;
-                    }
-                }
-                int num2;
-                num2 = Singleton<EconomyManager>.instance.GetTaxRate(@class, taxationPolicies);
-                if (MainDataStore.family_money[homeid] > 0)
-                {
-                    incomeAccumulation = (int)(num2 * incomeAccumulation * ((float)(instance.m_districts.m_buffer[district].GetLandValue() + 50) / 10000));
-                }
-                else
-                {
-                    incomeAccumulation = 0;
-                }
-            }
-
-            int educationFee = 0;
-            if ((Singleton<CitizenManager>.instance.m_citizens.m_buffer[citizenID].m_flags & Citizen.Flags.Student) != Citizen.Flags.None)
-            {
-                //Only university will cost money
-                bool isCampusDLC = false;
-                //Campus DLC cost 100
-                ushort visitBuilding = Singleton<CitizenManager>.instance.m_citizens.m_buffer[citizenID].m_visitBuilding;
-                if (visitBuilding != 0u)
-                {
-                    Building buildingData = Singleton<BuildingManager>.instance.m_buildings.m_buffer[visitBuilding];
-                    if (buildingData.Info.m_class.m_service == ItemClass.Service.PlayerEducation)
-                    {
-                        educationFee = 100;
-                        isCampusDLC = true;
-                    }
-                }
-
-                if (!isCampusDLC && (Singleton<CitizenManager>.instance.m_citizens.m_buffer[citizenID].m_flags.IsFlagSet(Citizen.Flags.Education2)))
-                {
-                    educationFee = 20;
-                    Singleton<EconomyManager>.instance.AddPrivateIncome(20, ItemClass.Service.Education, ItemClass.SubService.None, ItemClass.Level.Level3, 115333);
-                }
-            }
-            return educationFee;
-        }
-
-        public static void GetVoteTickets()
-        {
-            System.Random rand = new System.Random();
-            if (Politics.cPartyChance + Politics.gPartyChance + Politics.sPartyChance + Politics.lPartyChance + Politics.nPartyChance != (800 + RealCityEconomyExtension.partyTrendStrength))
-            {
-                if (rand.Next(64) <= 1)
-                {
-                    DebugLog.LogToFileOnly("Error: Chance is not equal 800 " + (Politics.cPartyChance + Politics.gPartyChance + Politics.sPartyChance + Politics.lPartyChance + Politics.nPartyChance).ToString());
-                }
-            }
-
-            if (RealCityEconomyExtension.voteRandom < Politics.cPartyChance)
-            {
-                Politics.cPartyTickets++;
-            }
-            else if (RealCityEconomyExtension.voteRandom < Politics.cPartyChance + Politics.gPartyChance)
-            {
-                Politics.gPartyTickets++;
-            }
-            else if (RealCityEconomyExtension.voteRandom < Politics.cPartyChance + Politics.gPartyChance + Politics.sPartyChance)
-            {
-                Politics.sPartyTickets++;
-            }
-            else if (RealCityEconomyExtension.voteRandom < Politics.cPartyChance + Politics.gPartyChance + Politics.sPartyChance + Politics.lPartyChance)
-            {
-                Politics.lPartyTickets++;
-            }
-            else
-            {
-                Politics.nPartyTickets++;
-            }
-        }
-
-        public static  void GetVoteChance(uint citizenID, Citizen citizen, uint homeID)
-        {
-            if ((int)Citizen.GetAgeGroup(citizen.m_age) >= 2)
-            {
-                if (Politics.parliamentCount == 1)
-                {
-                    Politics.cPartyChance = 0;
-                    Politics.gPartyChance = 0;
-                    Politics.sPartyChance = 0;
-                    Politics.lPartyChance = 0;
-                    Politics.nPartyChance = 0;
-
-                    Politics.cPartyChance += (ushort)(Politics.education[(int)citizen.EducationLevel, 0] << 1);
-                    Politics.gPartyChance += (ushort)(Politics.education[(int)citizen.EducationLevel, 1] << 1);
-                    Politics.sPartyChance += (ushort)(Politics.education[(int)citizen.EducationLevel, 2] << 1);
-                    Politics.lPartyChance += (ushort)(Politics.education[(int)citizen.EducationLevel, 3] << 1);
-                    Politics.nPartyChance += (ushort)(Politics.education[(int)citizen.EducationLevel, 4] << 1);
-
-                    int idex = 14;
-                    switch (Singleton<BuildingManager>.instance.m_buildings.m_buffer[citizen.m_workBuilding].Info.m_class.m_service)
-                    {
-                        case ItemClass.Service.Beautification:
-                        case ItemClass.Service.Road:
-                        case ItemClass.Service.Water:
-                        case ItemClass.Service.FireDepartment:
-                        case ItemClass.Service.PoliceDepartment:
-                        case ItemClass.Service.HealthCare:
-                        case ItemClass.Service.Garbage:
-                        case ItemClass.Service.PublicTransport:
-                        case ItemClass.Service.Disaster:
-                        case ItemClass.Service.Education:
-                        case ItemClass.Service.Electricity:
-                        case ItemClass.Service.Monument:
-                            idex = 0; break;
-                    }
-
-                    switch (Singleton<BuildingManager>.instance.m_buildings.m_buffer[citizen.m_workBuilding].Info.m_class.m_subService)
-                    {
-                        case ItemClass.SubService.CommercialLow:
-                        case ItemClass.SubService.CommercialHigh:
-                            if (Singleton<BuildingManager>.instance.m_buildings.m_buffer[citizen.m_workBuilding].Info.m_class.m_level == ItemClass.Level.Level1)
-                            {
-                                idex = 1;
-                            }
-                            else if (Singleton<BuildingManager>.instance.m_buildings.m_buffer[citizen.m_workBuilding].Info.m_class.m_level == ItemClass.Level.Level2)
-                            {
-                                idex = 2;
-                            }
-                            else if (Singleton<BuildingManager>.instance.m_buildings.m_buffer[citizen.m_workBuilding].Info.m_class.m_level == ItemClass.Level.Level3)
-                            {
-                                idex = 3;
-                            }
-                            break;
-                        case ItemClass.SubService.CommercialTourist:
-                        case ItemClass.SubService.CommercialLeisure:
-                            idex = 4; break;
-                        case ItemClass.SubService.CommercialEco:
-                            idex = 5; break;
-                        case ItemClass.SubService.IndustrialGeneric:
-                            if (Singleton<BuildingManager>.instance.m_buildings.m_buffer[citizen.m_workBuilding].Info.m_class.m_level == ItemClass.Level.Level1)
-                            {
-                                idex = 6;
-                            }
-                            else if (Singleton<BuildingManager>.instance.m_buildings.m_buffer[citizen.m_workBuilding].Info.m_class.m_level == ItemClass.Level.Level2)
-                            {
-                                idex = 7;
-                            }
-                            else if (Singleton<BuildingManager>.instance.m_buildings.m_buffer[citizen.m_workBuilding].Info.m_class.m_level == ItemClass.Level.Level3)
-                            {
-                                idex = 8;
-                            }
-                            break;
-                        case ItemClass.SubService.IndustrialFarming:
-                        case ItemClass.SubService.IndustrialForestry:
-                        case ItemClass.SubService.IndustrialOil:
-                        case ItemClass.SubService.IndustrialOre:
-                            idex = 9; break;
-                        case ItemClass.SubService.OfficeGeneric:
-                            if (Singleton<BuildingManager>.instance.m_buildings.m_buffer[citizen.m_workBuilding].Info.m_class.m_level == ItemClass.Level.Level1)
-                            {
-                                idex = 10;
-                            }
-                            else if (Singleton<BuildingManager>.instance.m_buildings.m_buffer[citizen.m_workBuilding].Info.m_class.m_level == ItemClass.Level.Level2)
-                            {
-                                idex = 11;
-                            }
-                            else if (Singleton<BuildingManager>.instance.m_buildings.m_buffer[citizen.m_workBuilding].Info.m_class.m_level == ItemClass.Level.Level3)
-                            {
-                                idex = 12;
-                            }
-                            break;
-                        case ItemClass.SubService.OfficeHightech:
-                            idex = 13; break;
-                    }
-
-                    if (idex < 0 || idex > 14)
-                    {
-                        DebugLog.LogToFileOnly("Error workplace idex" + idex.ToString());
-                    }
-
-
-                    Politics.cPartyChance += (ushort)(Politics.workplace[idex, 0] << 1);
-                    Politics.gPartyChance += (ushort)(Politics.workplace[idex, 1] << 1);
-                    Politics.sPartyChance += (ushort)(Politics.workplace[idex, 2] << 1);
-                    Politics.lPartyChance += (ushort)(Politics.workplace[idex, 3] << 1);
-                    Politics.nPartyChance += (ushort)(Politics.workplace[idex, 4] << 1);
-
-                    if (MainDataStore.family_money[homeID] < 5000)
-                    {
-                        idex = 0;
-                    }
-                    else if (MainDataStore.family_money[homeID] >= 20000)
-                    {
-                        idex = 2;
-                    }
-                    else
-                    {
-                        idex = 1;
-                    }
-
-                    if (idex < 0 || idex > 3)
-                    {
-                        DebugLog.LogToFileOnly("Error: Invaid money idex = " + idex.ToString());
-                    }
-                    Politics.cPartyChance += (ushort)(Politics.money[idex, 0] << 1);
-                    Politics.gPartyChance += (ushort)(Politics.money[idex, 1] << 1);
-                    Politics.sPartyChance += (ushort)(Politics.money[idex, 2] << 1);
-                    Politics.lPartyChance += (ushort)(Politics.money[idex, 3] << 1);
-                    Politics.nPartyChance += (ushort)(Politics.money[idex, 4] << 1);
-
-                    int temp = 0;
-
-                    temp = (int)Citizen.GetAgeGroup(citizen.m_age) - 2;
-
-                    if (temp < 0)
-                    {
-                        DebugLog.LogToFileOnly(temp.ToString() + Citizen.GetAgeGroup(citizen.m_age).ToString());
-                    }
-
-                    Politics.cPartyChance += Politics.age[temp, 0];
-                    Politics.gPartyChance += Politics.age[temp, 1];
-                    Politics.sPartyChance += Politics.age[temp, 2];
-                    Politics.lPartyChance += Politics.age[temp, 3];
-                    Politics.nPartyChance += Politics.age[temp, 4];
-
-                    temp = (int)Citizen.GetGender(citizenID);
-
-
-                    Politics.cPartyChance += Politics.gender[temp, 0];
-                    Politics.gPartyChance += Politics.gender[temp, 1];
-                    Politics.sPartyChance += Politics.gender[temp, 2];
-                    Politics.lPartyChance += Politics.gender[temp, 3];
-                    Politics.nPartyChance += Politics.gender[temp, 4];
-
-                    if (RealCityEconomyExtension.partyTrend == 0)
-                    {
-                        Politics.cPartyChance += RealCityEconomyExtension.partyTrendStrength;
-                    }
-                    else if (RealCityEconomyExtension.partyTrend == 1)
-                    {
-                        Politics.gPartyChance += RealCityEconomyExtension.partyTrendStrength;
-                    }
-                    else if (RealCityEconomyExtension.partyTrend == 2)
-                    {
-                        Politics.sPartyChance += RealCityEconomyExtension.partyTrendStrength;
-                    }
-                    else if (RealCityEconomyExtension.partyTrend == 3)
-                    {
-                        Politics.lPartyChance += RealCityEconomyExtension.partyTrendStrength;
-                    }
-                    else if (RealCityEconomyExtension.partyTrend == 4)
-                    {
-                        Politics.nPartyChance += RealCityEconomyExtension.partyTrendStrength;
-                    }
-                    else
-                    {
-                        DebugLog.LogToFileOnly("Error: Invalid partyTrend = " + RealCityEconomyExtension.partyTrend.ToString());
-                    }
-
-                    GetVoteTickets();
-                }
-            }
-        }
+        }     
     }
 }
