+++ Patch/CargoTruckAIReleaseVehiclePatch.cs
@@ -1,5 +1,5 @@
 ﻿using ColossalFramework;
-using Harmony;
+using HarmonyLib;
 using RealCity.CustomAI;
 using RealCity.Util;
 using System;
@@ -8,7 +8,7 @@ using System.Reflection;
 namespace RealCity.Patch
 {
     [HarmonyPatch]
-    public static class CargoTruckAIReleaseVehiclePatch
+    public class CargoTruckAIReleaseVehiclePatch
     {
         public static MethodBase TargetMethod()
         {
