+++ Patch/TouristAISimulationStepPatch.cs
@@ -1,5 +1,5 @@
 ﻿using ColossalFramework;
-using Harmony;
+using HarmonyLib;
 using RealCity.Util;
 using System;
 using System.Reflection;
@@ -7,7 +7,7 @@ using System.Reflection;
 namespace RealCity.Patch
 {
     [HarmonyPatch]
-    public static class TouristAISimulationStepPatch
+    public class TouristAISimulationStepPatch
     {
         public static MethodBase TargetMethod()
         {
