+++ Patch/VehicleManagerReleaseVehicleImplementationPatch.cs
@@ -1,4 +1,4 @@
-﻿using Harmony;
+﻿using HarmonyLib;
 using RealCity.Util;
 using System;
 using System.Reflection;
@@ -8,8 +8,6 @@ namespace RealCity.Patch
     [HarmonyPatch]
     public class VehicleManagerReleaseVehicleImplementationPatch
     {
-        public static byte[] watingPathTime = new byte[65536];
-        public static ushort[] stuckTime = new ushort[65536];
         public static MethodBase TargetMethod()
         {
             return typeof(VehicleManager).GetMethod("ReleaseVehicleImplementation", BindingFlags.NonPublic | BindingFlags.Instance, null, new Type[] { typeof(ushort), typeof(Vehicle).MakeByRefType() }, null);
@@ -18,8 +16,8 @@ namespace RealCity.Patch
         {
             MainDataStore.vehicleTransferTime[vehicle] = 0;
             MainDataStore.isVehicleCharged[vehicle] = false;
-            watingPathTime[vehicle] = 0;
-            stuckTime[vehicle] = 0;
+            MainDataStore.watingPathTime[vehicle] = 0;
+            MainDataStore.stuckTime[vehicle] = 0;
         }
     }
 }
