+++ RealCityThreading.cs
@@ -6,7 +6,7 @@ using ColossalFramework;
 using System;
 using System.Reflection;
 using RealCity.CustomAI;
-using RealCity.Patch;
+using HarmonyLib;
 
 namespace RealCity
 {
@@ -134,12 +134,13 @@ namespace RealCity
                 }
                 else
                 {
-                    var methods = HarmonyDetours.harmony.GetPatchedMethods();
+                    var harmony = new Harmony(HarmonyDetours.ID);
+                    var methods = harmony.GetPatchedMethods();
                     int i = 0;
                     foreach (var method in methods)
                     {
-                        var info = HarmonyDetours.harmony.GetPatchInfo(method);
-                        if (info.Owners?.Contains(HarmonyDetours.harmony.Id) == true)
+                        var info = Harmony.GetPatchInfo(method);
+                        if (info.Owners?.Contains(HarmonyDetours.ID) == true)
                         {
                             DebugLog.LogToFileOnly("Harmony patch method = " + method.Name.ToString());
                             if (info.Prefixes.Count != 0)
@@ -172,12 +173,12 @@ namespace RealCity
             {
                 if (vehicleData.m_flags.IsFlagSet(Vehicle.Flags.WaitingPath))
                 {
-                    VehicleManagerReleaseVehicleImplementationPatch.stuckTime[vehicleID] = 0;
+                    MainDataStore.stuckTime[vehicleID] = 0;
                     if (vehicleData.m_path != 0)
                     {
-                        VehicleManagerReleaseVehicleImplementationPatch.watingPathTime[vehicleID]++;
+                        MainDataStore.watingPathTime[vehicleID]++;
                     }
-                    if (VehicleManagerReleaseVehicleImplementationPatch.watingPathTime[vehicleID] > 10)
+                    if (MainDataStore.watingPathTime[vehicleID] > 10)
                     {
                         ushort building = 0;
                         if (!vehicleData.m_flags.IsFlagSet(Vehicle.Flags.GoingBack))
@@ -196,7 +197,7 @@ namespace RealCity
                         DebugLog.LogToFileOnly("DebugInfo: Stuck at waitingpath vehicle flag is " + vehicleData.m_flags.ToString());
                         DebugLog.LogToFileOnly("DebugInfo: Stuck at waitingpath vehicle ai is " + vehicleData.Info.m_vehicleAI.ToString());
                         DebugLog.LogToFileOnly("DebugInfo: Stuck at waitingpath vehicle pathflag is " + Singleton<PathManager>.instance.m_pathUnits.m_buffer[vehicleData.m_path].m_simulationFlags.ToString());
-                        VehicleManagerReleaseVehicleImplementationPatch.watingPathTime[vehicleID] = 0;
+                        MainDataStore.watingPathTime[vehicleID] = 0;
                         Singleton<PathManager>.instance.ReleasePath(vehicleData.m_path);
                         vehicleData.m_path = 0u;
                         vehicleData.m_flags = (vehicleData.m_flags & ~Vehicle.Flags.WaitingPath);
@@ -204,25 +205,25 @@ namespace RealCity
                 }
                 else
                 {
-                    VehicleManagerReleaseVehicleImplementationPatch.watingPathTime[vehicleID] = 0;
+                    MainDataStore.watingPathTime[vehicleID] = 0;
                     if (!vehicleData.m_flags.IsFlagSet(Vehicle.Flags.WaitingCargo) && !vehicleData.m_flags.IsFlagSet(Vehicle.Flags.WaitingSpace) && !vehicleData.m_flags.IsFlagSet(Vehicle.Flags.WaitingLoading) && !vehicleData.m_flags.IsFlagSet(Vehicle.Flags.WaitingTarget) && !vehicleData.m_flags.IsFlagSet(Vehicle.Flags.WaitingSpace) && !vehicleData.m_flags.IsFlagSet(Vehicle.Flags.Stopped) && !vehicleData.m_flags.IsFlagSet(Vehicle.Flags.Congestion))
                     {
                         float realSpeed = (float)Math.Sqrt(vehicleData.GetLastFrameVelocity().x * vehicleData.GetLastFrameVelocity().x + vehicleData.GetLastFrameVelocity().y * vehicleData.GetLastFrameVelocity().y + vehicleData.GetLastFrameVelocity().z * vehicleData.GetLastFrameVelocity().z);
                         if (realSpeed < 0.1f)
                         {
-                            VehicleManagerReleaseVehicleImplementationPatch.stuckTime[vehicleID]++;
+                            MainDataStore.stuckTime[vehicleID]++;
                         }
                         else
                         {
-                            VehicleManagerReleaseVehicleImplementationPatch.stuckTime[vehicleID] = 0;
+                            MainDataStore.stuckTime[vehicleID] = 0;
                         }
 
-                        if (VehicleManagerReleaseVehicleImplementationPatch.stuckTime[vehicleID] > 600)
+                        if (MainDataStore.stuckTime[vehicleID] > 600)
                         {
                             DebugLog.LogToFileOnly("DebugInfo: Stuck vehicle transfer type is " + vehicleData.m_transferType.ToString());
                             DebugLog.LogToFileOnly("DebugInfo: Stuck vehicle flag is " + vehicleData.m_flags.ToString());
                             DebugLog.LogToFileOnly("DebugInfo: Stuck vehicle ai is " + vehicleData.Info.m_vehicleAI.ToString());
-                            VehicleManagerReleaseVehicleImplementationPatch.stuckTime[vehicleID] = 0;
+                            MainDataStore.stuckTime[vehicleID] = 0;
                             Singleton<PathManager>.instance.ReleasePath(vehicleData.m_path);
                             vehicleData.m_path = 0u;
                             Singleton<VehicleManager>.instance.ReleaseVehicle(vehicleID);
