+++ RebalancedIndustries/RI_Detours.cs
@@ -5,7 +5,7 @@ using ColossalFramework;
 using ColossalFramework.Globalization;
 using ColossalFramework.UI;
 using System.Reflection;
-using Harmony;
+using HarmonyLib;
 
 namespace RealCity.RebalancedIndustries
 {
